refactor: extract Contact methods to separate Contacts class
Contact系のメソッドをPhysBoneExtractor.csから新しいContacts.csファイルに分離:
- CopyVRCContacts()メソッドの移動
- CopyVRCContactSender()およびCopyVRCContactReceiver()メソッドの移動
- PhysBoneExtractor.csでContactsクラスを使用するよう修正
- コードの責任分離とメンテナビリティの向上
