+++ Editor/PhysBoneExtractor.cs
@@ -187,7 +187,8 @@ namespace dev.kesera2.physbone_extractor
                         copiedPbComponents =
                             Utility.CopyComponent<VRCPhysBone>(searchRoot, _avatarDynamics, pbGameObjectName,
                                 CopyVRCPhysBone);
-                        var hasContacts = CopyVRCContacts();
+                        var contacts = new Contacts(contactsGameObjectName, contactSenderGameObjectName, contactReceiverGameObjectName, splitContacts);
+                        var hasContacts = contacts.CopyVRCContacts(searchRoot, _avatarDynamics, out copiedContactSenderComponents, out copiedContactReceiverComponents);
                         if (!(copiedPbComponents.Count > 0 || copiedPbColliderComponents.Count > 0 || hasContacts))
                             // Destroy the GameObject of AvatarDynamics if there is no target components.
                             DestroyImmediate(_avatarDynamics);
@@ -297,26 +298,6 @@ namespace dev.kesera2.physbone_extractor
         }
 
 
-        private bool CopyVRCContacts()
-        {
-            var contactsParent = new GameObject(contactsGameObjectName);
-            contactsParent.transform.SetParent(_avatarDynamics.transform);
-
-            copiedContactReceiverComponents = Utility.CopyComponent<VRCContactReceiver>(searchRoot, contactsParent,
-                contactReceiverGameObjectName,
-                CopyVRCContactReceiver, !splitContacts);
-            copiedContactSenderComponents = Utility.CopyComponent<VRCContactSender>(searchRoot, contactsParent,
-                contactSenderGameObjectName,
-                CopyVRCContactSender, !splitContacts);
-
-            if (copiedContactReceiverComponents.Count == 0 && copiedContactSenderComponents.Count == 0)
-            {
-                DestroyImmediate(contactsParent);
-                return false;
-            }
-
-            return true;
-        }
 
 
         private void CopyVRCPhysBone(VRCPhysBone source, VRCPhysBone destination)
@@ -424,36 +405,6 @@ namespace dev.kesera2.physbone_extractor
                 .FirstOrDefault(t => t.name.ToLower() == "armature");
         }
 
-        private void CopyVRCContactSender(VRCContactSender source, VRCContactSender destination)
-        {
-            destination.name = source.name;
-            destination.rootTransform = source.rootTransform;
-            destination.shapeType = source.shapeType;
-            destination.radius = source.radius;
-            destination.position = source.position;
-            destination.rotation = source.rotation;
-            destination.collisionTags = source.collisionTags;
-        }
-
-        private void CopyVRCContactReceiver(VRCContactReceiver source, VRCContactReceiver destination)
-        {
-            destination.name = source.name;
-            destination.rootTransform = source.rootTransform;
-            // Shape
-            destination.shapeType = source.shapeType;
-            destination.radius = source.radius;
-            destination.position = source.position;
-            destination.rotation = source.rotation;
-            // Filtering
-            destination.allowSelf = source.allowSelf;
-            destination.allowOthers = source.allowOthers;
-            destination.localOnly = source.localOnly;
-            destination.collisionTags = source.collisionTags;
-            // Receiver
-            destination.receiverType = source.receiverType;
-            destination.parameter = source.parameter;
-            destination.minVelocity = source.minVelocity;
-        }
 
         private void RemoveOriginalComponent()
         {
